requirements.txt
requests
python-dotenv
requests-toolbelt